+++ app/about/page.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+const About = () => {
+  return <div>About</div>;
+};
+
+export default About;
