+++ app/home/[slug]/page.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function SubHome({ params }: { params: { slug: string } }) {
+  return <p>Route ID: {params.slug}</p>;
+}
