+++ app/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 import React from "react";
-import HomePage from "./pages/home/page";
+import HomePage from "./home/page";
 import Footer from "./components/organisms/footer";
 import NavBar from "./components/organisms/navbar";
