--- app/pages/about/page.tsx
@@ -1,3 +1,0 @@
-export default function About() {
-  return <div>About</div>;
-}
