--- app/pages/home/page.tsx
@@ -1,18 +1,0 @@
-"use client";
-import Footer from "@/app/components/organisms/footer";
-import React from "react";
-import styled from "styled-components";
-
-const Container = styled.div`
-  height: 80vh;
-  background-color: #ffff;
-`;
-const HomePage = () => {
-  return (
-    <Container>
-      <h1>Home</h1>
-    </Container>
-  );
-};
-
-export default HomePage;
